+++ .kotlin/errors/errors-1730230416781.log
@@ -1,0 +1,4 @@
+kotlin version: 2.1.0-Beta2
+error message: The daemon has terminated unexpectedly on startup attempt #1 with error code: Unknown. The daemon process output:
+    1. Kotlin compile daemon is ready
+
