+++ gradle/libs.versions.toml
@@ -14,26 +14,26 @@ composeMaterial = "1.7.4"
 composeMaterial3 = "1.3.0"
 composeUi = "1.7.4"
 composeUiTest = "1.7.4"
+foundation = "1.7.4"
 constraintLayout = "2.1.4"
 coreKtx = "1.13.1"
 customView = "1.2.0-alpha02"
-daggerHilt = "2.52"
+daggerHilt = "2.51.1" #don't change
 dokka = "1.9.20"
 espresso = "3.6.1"
 firebase = "33.5.1"
 flowLayout = "0.31.1"
-foundation = "1.7.4"
 gradleVersionsPlugin = "0.47.0"
 hiyaJacoco = "0.3"
 jacoco = "0.8.10"
 junit = "4.13.2"
 kakao = "3.4.3"
 klint = "0.50.0"
-kotlinCompiler = "1.5.3"
+kotlinCompiler = "1.5.15"
 kotlinComposeVersion = "2.0.21"
 kotlinHilt = "1.2.0"
-kotlinKapt = "2.1.0-Beta2"
-kotlinVersion = "2.1.0-Beta2"
+kotlinKapt = "1.9.25" #don't change
+kotlinVersion = "1.9.25" #don'tchange
 kotlinxCoroutines = "1.9.0"
 kotlinxMetadataJvm = "0.9.0"
 kotlinxSerializationJson = "1.7.3"
@@ -59,6 +59,7 @@ spotless = "6.25.0"
 timber = "5.0.1"
 truth = "1.1.5"
 vanniktechMavenPublish = "0.25.3"
+workRuntimeKtx = "2.9.1"
 
 [libraries]
 agoda-kakao = { group = "io.github.kakaocup", name = "kakao", version.ref = "kakao" }
@@ -96,6 +97,7 @@ androidx-lifecycle-service = { group = "androidx.lifecycle", name = "lifecycle-s
 androidx-lifecycle-viewmodel-compose = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-compose", version.ref = "lifecycle" }
 androidx-lifecycle-viewmodel-ktx = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-ktx", version.ref = "lifecycle" }
 androidx-lifecycle-viewmodel-savedstate = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-savedstate", version.ref = "lifecycle" }
+androidx-lifecycle-viewmodel-android = { group = "androidx.lifecycle", name = "lifecycle-viewmodel-android", version.ref = "lifecycle" }
 androidx-navigation-compose = { group = "androidx.navigation", name = "navigation-compose", version.ref = "navigation" }
 androidx-navigation-dynamic = { group = "androidx.navigation", name = "navigation-dynamic-features-fragment", version.ref = "navigation" }
 androidx-navigation-fragment = { group = "androidx.navigation", name = "navigation-fragment-ktx", version.ref = "navigation" }
@@ -108,7 +110,7 @@ androidx-test-rules = { group = "androidx.test", name = "rules", version.ref = "
 androidx-test-runner = { group = "androidx.test", name = "runner", version.ref = "runner" }
 chuckerteam-chucker = { group = "com.github.chuckerteam.chucker", name = "library", version.ref = "chucker" }
 chuckerteam-chucker-noop = { group = "com.github.chuckerteam.chucker", name = "library-no-op", version.ref = "chucker" }
-compose-compiler = { group = "androidx.compose.compiler", name ="*", version.ref = "kotlinCompiler" }
+compose-compiler = { group = "androidx.compose.compiler", name ="compiler", version.ref = "kotlinCompiler" }
 google-accompanist-flowlayout = { group = "com.google.accompanist", name = "accompanist-flowlayout", version.ref = "flowLayout" }
 google-android-material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 google-dagger-hilt = { group = "com.google.dagger", name = "hilt-android", version.ref = "daggerHilt" }
@@ -136,6 +138,7 @@ squareup-okhttp-logging = { group = "com.squareup.okhttp3", name = "logging-inte
 squareup-okhttp-mockwebserver = { group = "com.squareup.okhttp3", name = "mockwebserver", version.ref = "mockWebServer" }
 squareup-retrofit = { group = "com.squareup.retrofit2", name = "retrofit", version.ref = "retrofit" }
 squareup-retrofit-gson = { group = "com.squareup.retrofit2", name = "converter-gson", version.ref = "retrofit" }
+work-runtime-ktx = { group = "androidx.work", name = "work-runtime-ktx", version.ref = "workRuntimeKtx" }
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "application" }
