+++ sample/src/main/res/navigation/navigation_graph.xml
@@ -2,14 +2,14 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/navigation_graph"
-    app:startDestination="@id/mainScreenFragment"
+    app:startDestination="@id/homeScreenFragment"
     app:defaultNavHost="true"
     app:navGraph="@navigation/navigation_graph"
     >
 
     <fragment
-        android:id="@+id/mainScreenFragment"
-        android:name="io.rekast.sdk.sample.ui.mainscreen.MainScreenFragment"
+        android:id="@+id/homeScreenFragment"
+        android:name="io.rekast.sdk.sample.views.home.HomeScreenFragment"
         android:label="@string/home_screen">
         <action
             android:id="@+id/navigateToCollectionPay"
@@ -29,11 +29,11 @@
     </fragment>
     <fragment
         android:id="@+id/collectionPayScreenFragment"
-        android:name="io.rekast.sdk.sample.ui.collection.pay.CollectionPayScreenFragment"
+        android:name="io.rekast.sdk.sample.views.collection.pay.CollectionPayScreenFragment"
         android:label="@string/collections_pay_screen">
         <action
             android:id="@+id/navigateToHome"
-            app:destination="@+id/mainScreenFragment" />
+            app:destination="@+id/homeFragment" />
         <action
             android:id="@+id/navigateToCollectionWithdraw"
             app:destination="@+id/collectionWithDrawScreenFragment" />
@@ -50,11 +50,11 @@
 
     <fragment
         android:id="@+id/collectionWithDrawScreenFragment"
-        android:name="io.rekast.sdk.sample.ui.collection.withdraw.CollectionWithdrawScreenFragment"
+        android:name="io.rekast.sdk.sample.views.collection.withdraw.CollectionWithdrawScreenFragment"
         android:label="@string/collections_withdraw_screen">
         <action
             android:id="@+id/navigateToHome"
-            app:destination="@+id/mainScreenFragment" />
+            app:destination="@+id/homeFragment" />
         <action
             android:id="@+id/navigateToCollectionPay"
             app:destination="@+id/collectionPayScreenFragment" />
@@ -71,11 +71,11 @@
 
     <fragment
         android:id="@+id/disbursementDepositScreenFragment"
-        android:name="io.rekast.sdk.sample.ui.disbursement.deposit.DisbursementDepositScreenFragment"
+        android:name="io.rekast.sdk.sample.views.disbursement.deposit.DisbursementDepositScreenFragment"
         android:label="@string/disbursement_deposit_screen">
         <action
             android:id="@+id/navigateToHome"
-            app:destination="@+id/mainScreenFragment" />
+            app:destination="@+id/homeFragment" />
         <action
             android:id="@+id/navigateToCollectionPay"
             app:destination="@+id/collectionPayScreenFragment" />
@@ -92,11 +92,11 @@
 
     <fragment
         android:id="@+id/disbursementRefundScreenFragment"
-        android:name="io.rekast.sdk.sample.ui.disbursement.refund.DisbursementRefundScreenFragment"
+        android:name="io.rekast.sdk.sample.views.disbursement.refund.DisbursementRefundScreenFragment"
         android:label="@string/disbursement_Refund_screen">
         <action
             android:id="@+id/navigateToHome"
-            app:destination="@+id/mainScreenFragment" />
+            app:destination="@+id/homeFragment" />
         <action
             android:id="@+id/navigateToCollectionPay"
             app:destination="@+id/collectionPayScreenFragment" />
@@ -113,12 +113,12 @@
 
     <fragment
         android:id="@+id/remittanceScreenFragment"
-        android:name="io.rekast.sdk.sample.ui.remittance.RemittanceScreenFragment"
+        android:name="io.rekast.sdk.sample.views.remittance.RemittanceScreenFragment"
         android:label="@string/remittance_screen">
 
         <action
             android:id="@+id/navigateToHome"
-            app:destination="@+id/mainScreenFragment" />
+            app:destination="@+id/homeFragment" />
         <action
             android:id="@+id/navigateToCollectionPay"
             app:destination="@+id/collectionPayScreenFragment" />
